Refactor table.py into seperate modules for better maintainability
