+++ netbox_librenms_plugin/tables/interfaces.py
@@ -1,14 +1,14 @@
 import django_tables2 as tables
-from django.middleware.csrf import get_token
+from django.core.cache import cache
 from django.utils.html import format_html
 from django.utils.safestring import mark_safe
-from netbox.tables import NetBoxTable, columns
 from netbox.tables.columns import BooleanColumn, ToggleColumn
-from utilities.paginator import EnhancedPaginator, get_paginate_count
+from utilities.paginator import EnhancedPaginator
 from utilities.templatetags.helpers import humanize_speed
+from functools import lru_cache
 
-from .models import InterfaceTypeMapping
-from .utils import (
+from netbox_librenms_plugin.models import InterfaceTypeMapping
+from netbox_librenms_plugin.utils import (
     convert_speed_to_kbps,
     format_mac_address,
     get_interface_name_field,
@@ -358,7 +358,7 @@ class VCInterfaceTable(LibreNMSInterfaceTable):
             '<select name="device_selection_{0}" id="{1}" class="form-select vc-member-select" data-interface="{0}" data-row-id="{0}">{2}</select>',
             interface_name,
             base_id,
-            mark_safe("".join(options))
+            mark_safe("".join(options)),
         )
 
     def format_interface_data(self, port_data, device):
@@ -386,107 +386,6 @@ class VCInterfaceTable(LibreNMSInterfaceTable):
         }
 
 
-class InterfaceTypeMappingTable(NetBoxTable):
-    """
-    Table for displaying InterfaceTypeMapping data.
-    """
-
-    librenms_type = tables.Column(verbose_name="LibreNMS Type")
-    librenms_speed = tables.Column(verbose_name="LibreNMS Speed (Kbps)")
-    netbox_type = tables.Column(verbose_name="NetBox Type")
-    actions = columns.ActionsColumn(actions=("edit", "delete"))
-
-    class Meta:
-        model = InterfaceTypeMapping
-        fields = ("id", "librenms_type", "librenms_speed", "netbox_type", "actions")
-        default_columns = (
-            "id",
-            "librenms_type",
-            "librenms_speed",
-            "netbox_type",
-            "actions",
-        )
-        attrs = {"class": "table table-hover table-headings table-striped"}
-
-
-class SiteLocationSyncTable(tables.Table):
-    """
-    Table for displaying Netbox Site and Librenms Location data.
-    """
-
-    netbox_site = tables.Column(linkify=True)
-    latitude = tables.Column(accessor="netbox_site.latitude")
-    longitude = tables.Column(accessor="netbox_site.longitude")
-    librenms_location = tables.Column(
-        accessor="librenms_location.location", verbose_name="LibreNMS Location"
-    )
-    librenms_latitude = tables.Column(
-        accessor="librenms_location.lat", verbose_name="LibreNMS Latitude"
-    )
-    librenms_longitude = tables.Column(
-        accessor="librenms_location.lng", verbose_name="LibreNMS Longitude"
-    )
-    actions = tables.Column(empty_values=())
-
-    def render_latitude(self, value, record):
-        return self.render_coordinate(value, record.is_synced)
-
-    def render_longitude(self, value, record):
-        return self.render_coordinate(value, record.is_synced)
-
-    def render_coordinate(self, value, is_synced):
-        css_class = "text-success" if is_synced else "text-danger"
-        return format_html('<span class="{}">{}</span>', css_class, value)
-
-    def render_actions(self, record):
-        csrf_token = get_token(self.request)
-        if record.is_synced:
-            return mark_safe(
-                '<span class="text-success"><i class="mdi mdi-check-circle" aria-hidden="true"></i> Synced</span>'
-            )
-        if record.librenms_location:
-            return mark_safe(
-                f'<form method="post">'
-                f'<input type="hidden" name="csrfmiddlewaretoken" value="{csrf_token}">'
-                f'<input type="hidden" name="action" value="update">'
-                f'<input type="hidden" name="pk" value="{record.netbox_site.pk}">'
-                '<button type="submit" class="btn btn-sm btn-warning">'
-                '<i class="mdi mdi-pencil" aria-hidden="true"></i> Update in LibreNMS'
-                "</button>"
-                "</form>"
-            )
-        else:
-            return mark_safe(
-                f'<form method="post">'
-                f'<input type="hidden" name="csrfmiddlewaretoken" value="{csrf_token}">'
-                f'<input type="hidden" name="action" value="create">'
-                f'<input type="hidden" name="pk" value="{record.netbox_site.pk}">'
-                '<button type="submit" class="btn btn-sm btn-primary">'
-                '<i class="mdi mdi-plus-thick" aria-hidden="true"></i> Create in LibreNMS'
-                "</button>"
-                "</form>"
-            )
-
-    def configure(self, request):
-        paginate = {
-            "paginator_class": EnhancedPaginator,
-            "per_page": get_paginate_count(request),
-        }
-        tables.RequestConfig(request, paginate).configure(self)
-
-    class Meta:
-        fields = (
-            "netbox_site",
-            "latitude",
-            "longitude",
-            "librenms_location",
-            "librenms_latitude",
-            "librenms_longitude",
-            "actions",
-        )
-        attrs = {"class": "table table-hover table-headings table-striped"}
-
-
 class LibreNMSVMInterfaceTable(LibreNMSInterfaceTable):
     """
     Table for displaying LibreNMS VM interface data.
@@ -509,147 +408,3 @@ class LibreNMSVMInterfaceTable(LibreNMSInterfaceTable):
     # Remove the type and speed column for VMs
     type = None
     speed = None
-
-
-class LibreNMSCableTable(tables.Table):
-    """
-    Table for displaying LibreNMS cable data.
-    """
-
-    selection = ToggleColumn(
-        accessor="local_port",
-        orderable=False,
-        visible=True,
-        attrs={"td": {"data-col": "selection"}, "input": {"name": "select"}},
-    )
-
-    local_port = tables.Column(
-        verbose_name="Local Port", attrs={"td": {"data-col": "local_port"}}
-    )
-    remote_port = tables.Column(
-        accessor="remote_port_name",
-        verbose_name="Remote Port",
-        attrs={"td": {"data-col": "remote_port"}},
-    )
-    remote_device = tables.Column(
-        verbose_name="Remote Device", attrs={"td": {"data-col": "remote_device"}}
-    )
-    cable_status = tables.Column(
-        verbose_name="Cable Status", attrs={"td": {"data-col": "cable_status"}}
-    )
-    actions = tables.TemplateColumn(
-        template_code="""
-        {% if record.can_create_cable %}
-            <button type="submit"
-                    class="btn btn-sm btn-primary"
-                    onclick="document.getElementById('selected_port').value='{{ record.local_port }}'">
-                Sync Cable
-            </button>
-        {% endif %}
-        """,
-        verbose_name="",
-        orderable=False,
-        attrs={"td": {"data-col": "actions"}},
-    )
-
-    def __init__(self, *args, device=None, **kwargs):
-        self.device = device
-        super().__init__(*args, **kwargs)
-        self.tab = "cables"
-        self.htmx_url = None
-        self.prefix = "cables_"
-
-    def render_remote_device(self, value, record):
-        if url := record.get("remote_device_url"):
-            return format_html('<a href="{}">{}</a>', url, value)
-        return value
-
-    def render_local_port(self, value, record):
-        if url := record.get("local_port_url"):
-            return format_html('<a href="{}">{}</a>', url, value)
-        return value
-
-    def render_remote_port(self, value, record):
-        if url := record.get("remote_port_url"):
-            return format_html('<a href="{}">{}</a>', url, value)
-        return value
-
-    def render_cable_status(self, value, record):
-        if url := record.get("cable_url"):
-            return format_html('<a href="{}">{}</a>', url, value)
-        return value
-
-    def configure(self, request):
-        paginate = {
-            "paginator_class": EnhancedPaginator,
-            "per_page": get_table_paginate_count(request, self.prefix),
-        }
-        tables.RequestConfig(request, paginate).configure(self)
-
-    class Meta:
-        sequence = [
-            "selection",
-            "local_port",
-            "remote_port",
-            "remote_device",
-            "cable_status",
-            "actions",
-        ]
-        row_attrs = {
-            "data-interface": lambda record: record["local_port"],
-            "data-device": lambda record: record["device_id"],
-            "data-name": lambda record: record["local_port"],
-        }
-        attrs = {"class": "table table-hover object-list", "id": "librenms-cable-table"}
-
-
-class VCCableTable(LibreNMSCableTable):
-    """
-    Table for displaying LibreNMS cable data for Virtual Chassis devices.
-    """
-
-    device_selection = tables.Column(
-        verbose_name="Virtual Chassis Member",
-        accessor="local_port",
-        attrs={"td": {"class": "device-selection-col", "data-col": "device_selection"}},
-    )
-
-    def __init__(self, *args, device=None, **kwargs):
-        super().__init__(*args, device=device, **kwargs)
-
-    def render_device_selection(self, value, record):
-        members = self.device.virtual_chassis.members.all()
-        chassis_member = get_virtual_chassis_member(self.device, record["local_port"])
-        selected_member_id = chassis_member.id if chassis_member else self.device.id
-
-        options = [
-            f'<option value="{member.id}"{" selected" if member.id == selected_member_id else ""}>{member.name}</option>'
-            for member in members
-        ]
-
-        return format_html(
-            '<select name="device_selection_{0}" id="device_selection_{0}" class="form-select" data-interface="{0}" data-row-id="{0}">{1}</select>',
-            record["local_port"],
-            mark_safe("".join(options)),
-        )
-
-    class Meta(LibreNMSCableTable.Meta):
-        sequence = [
-            "selection",
-            "device_selection",
-            "local_port",
-            "remote_port",
-            "remote_device",
-            "cable_status",
-            "actions",
-        ]
-        row_attrs = {
-            "data-interface": lambda record: record["local_port"],
-            "data-device": lambda record: record["device_id"],
-            "data-name": lambda record: record["local_port"],
-            "id": lambda record: record["local_port"],
-        }
-        attrs = {
-            "class": "table table-hover object-list",
-            "id": "librenms-cable-table-vc",
-        }
